+++ src/AdminPackage.php
@@ -49,6 +49,9 @@ final class AdminPackage extends Package
             'resources/admin.css' => 'css/admin.css',
             'resources/admin.tables.js' => 'js/admin.tables.js',
             'resources/admin.modal.js' => 'js/admin.modal.js',
+            'resources/js/admin.theme.js' => 'js/admin.theme.js',
+            'resources/js/admin.shell.js' => 'js/admin.shell.js',
+            'resources/img/vortexadmin.svg' => 'img/vortexadmin.svg',
         ];
     }